true/false voor single storage vervangen door checkbox

+++ eindwerkJava2/src/main/java/com/example/eindwerkJava2/repositories/LocationRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 
 public interface LocationRepository extends JpaRepository<Location, Long> {
 
-List<Location> findByLocationId(int locationId);
+List<Location> findByLocationId(Long locationId);
 
 }
